Attack on Merche (#323)
* Opcion validator/Isaarrveg (#321)

pcion validator que por algún motivo no se aplica

* Se pueden cambiar las autoridades. Restricciones del bloque 3 terminadas

+++ src/main/java/org/springframework/samples/petclinic/web/ComentarioController.java
@@ -34,7 +34,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class ComentarioController {
 	public static final String COMENTARIOS_FORM = "comentarios/createOrUpdateComentariosForm";
 	public static final String COMENTARIOS_LISTING = "comentarios/ComentariosListing";
+	public static final String MEJORAR_CUENTA = "usuarios/mejorarCuenta";
 	public static final String LOGIN = "login";
+	public static final String ERROR = "";
 
 	@Autowired
 	ComentarioService comentarioService;
@@ -55,8 +57,6 @@ public class ComentarioController {
 	private String auxViewHilo(int id, ModelMap model) {
 		Hilo hilo = hiloService.findById(id);
 		Collection<Comentario> comentarios = comentarioService.findByHiloId(id);
-		model.addAttribute("hilo", hilo);
-		model.addAttribute("comentarios", comentarios);
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		String username = authentication.getName();
 		Usuario usuario = usuarioService.findByUsername(username);
@@ -71,15 +71,22 @@ public class ComentarioController {
 				
 			}
 		}
+		model.addAttribute("hilo", hilo);
+		model.addAttribute("comentarios", comentarios);
+		model.addAttribute("usuario", usuario);
+		String authority = AuthController.highestLevel();
+		model.addAttribute("authority", authority);
 		return COMENTARIOS_LISTING;
 	}
 
 	@GetMapping("/{value}")
 	public String viewHilo(@PathVariable("value") int id, ModelMap model) {
-		if (!AuthController.hasPaid()) {
+		if (!AuthController.isAuthenticated()) {
 			return "redirect:/" + LOGIN;
 		}
-		System.out.println(AuthController.isAuthenticated());
+		if (!AuthController.hasPaid()) {
+			return "redirect:/" + MEJORAR_CUENTA;
+		}
 		return auxViewHilo(id, model);	
 	}
 
@@ -87,6 +94,9 @@ public class ComentarioController {
 	public String editNewComentario(ModelMap model, @PathVariable("value") int id) {
 		if (!AuthController.isAuthenticated()) {
 			return "redirect:/" + LOGIN;
+		}
+		if (!AuthController.hasPaid()) {
+			return "redirect:/" + MEJORAR_CUENTA;
 		}
 		Hilo hilo = hiloService.findById(id);
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -103,6 +113,9 @@ public class ComentarioController {
 			@PathVariable("value") int id, @PathVariable("cita") int cita) {
 		if (!AuthController.isAuthenticated()) {
 			return "redirect:/" + LOGIN;
+		}
+		if (!AuthController.hasPaid()) {
+			return "redirect:/" + MEJORAR_CUENTA;
 		}
 		Hilo hilo = hiloService.findById(id);
 		Collection<Usuario> usuarios = usuarioService.findAll();
@@ -121,7 +134,16 @@ public class ComentarioController {
 		if (!AuthController.isAuthenticated()) {
 			return "redirect:/" + LOGIN;
 		}
+		if (!AuthController.hasPaid()) {
+			return "redirect:/" + MEJORAR_CUENTA;
+		}
 		Comentario comentario = comentarioService.findById(comment);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		String username = authentication.getName();
+		Usuario usuarioLoggeado = usuarioService.findByUsername(username);
+		if (!comentario.getUsuario().equals(usuarioLoggeado) && !AuthController.isAdmin()) {
+			return "redirect:/" + ERROR;
+		}
 		comentarioService.delete(comentario);
 		model.addAttribute("message", "El comentario ha sido eliminado");
 		return viewHilo(id, model);
@@ -130,9 +152,6 @@ public class ComentarioController {
 	@PostMapping("/{value}/new")
 	public String saveNewComentario(@PathVariable("value") int id, @Valid Comentario comentario, BindingResult binding,
 			ModelMap model) {
-		if (!AuthController.isAuthenticated()) {
-			return "redirect:/" + LOGIN;
-		}
 		if (binding.hasErrors()) {
 			Collection<Usuario> usuarios = usuarioService.findAll();
 			model.addAttribute("usuarios", usuarios);
@@ -148,7 +167,7 @@ public class ComentarioController {
 	public String saveNewComentarioConCita(@PathVariable("value") int id, @Valid Comentario comentario, BindingResult binding,
 			ModelMap model) {
 		if (!AuthController.isAuthenticated()) {
-			return "redirect:/" + LOGIN;
+			return "redirect:/" + MEJORAR_CUENTA;
 		}
 
 
@@ -168,8 +187,17 @@ public class ComentarioController {
 		if (!AuthController.isAuthenticated()) {
 			return "redirect:/" + LOGIN;
 		}
-		Hilo hilo = hiloService.findById(value);
+		if (!AuthController.hasPaid()) {
+			return "redirect:/" + MEJORAR_CUENTA;
+		}
 		Comentario comentario = comentarioService.findById(comment);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		String username = authentication.getName();
+		Usuario usuarioLoggeado = usuarioService.findByUsername(username);
+		if (!comentario.getUsuario().equals(usuarioLoggeado) && !AuthController.isAdmin()) {
+			return "redirect:/" + ERROR;
+		}
+		Hilo hilo = hiloService.findById(value);
 		Collection<Usuario> usuarios = usuarioService.findAll();
 		model.addAttribute("hilo", hilo);
 		model.addAttribute("usuarios", usuarios);
@@ -180,9 +208,6 @@ public class ComentarioController {
 	@PostMapping("/{value}/edit/{id}")
 	public String editComentario(@PathVariable("id") int id, @Valid Comentario modifiedComentario,
 			BindingResult binding, ModelMap model) {
-		if (!AuthController.isAuthenticated()) {
-			return "redirect:/" + LOGIN;
-		}
 		Comentario comentario = comentarioService.findById(id);
 		if (binding.hasErrors()) {
 			Collection<Usuario> usuarios = usuarioService.findAll();
